refactor: refactored validation
